+++ gr-digital/lib/pfb_clock_sync_ccf_impl.cc
@@ -53,8 +53,8 @@ namespace gr {
 				     osps));
     }
 
-    static int ios[] = {sizeof(gr_complex), sizeof(float), sizeof(float), sizeof(float)};
-    static std::vector<int> iosig(ios, ios+sizeof(ios)/sizeof(int));
+    static size_t ios[] = {sizeof(gr_complex), sizeof(float), sizeof(float), sizeof(float)};
+    static std::vector<size_t> iosig(ios, ios+sizeof(ios)/sizeof(size_t));
     pfb_clock_sync_ccf_impl::pfb_clock_sync_ccf_impl(double sps, float loop_bw,
 						     const std::vector<float> &taps,
 						     unsigned int filter_size,
